requirements.txt
flask==2.2.2
cryptography==38.0.3
requests==2.28.1
gunicorn==20.1.0
python-dotenv==0.21.0